feat: introduce support to useWorkers
